feat(codegen): migrate io + evolution + tessellate + projection (24 methods)
Delete 4 more hand-written files (888 lines total):
- io.cpp (175 lines, 6 methods): importStep, exportStep, importStl, exportStl, toBREP, fromBREP
- evolution.cpp (325 lines, 12 methods): all *WithHistory methods
- tessellate.cpp (332 lines, 5 methods): tessellate, meshShape, meshBatch, wireframe, hasTriangulation
- projection.cpp (56 lines, 1 method): projectEdges

Add 5 struct ReturnType variants: EvolutionData, MeshData, MeshBatchData, EdgeData, ProjectionData.

149 methods now in codegen specs. Only kernel.cpp (3), xcaf.cpp (12), and bindings.cpp remain hand-written. All 212 tests pass.
