--- facade/src/tessellate.cpp
@@ -1,332 +1,0 @@
-#include "occt_kernel.h"
-
-#include <BRepAdaptor_Curve.hxx>
-#include <BRepLib_ToolTriangulatedShape.hxx>
-#include <BRepMesh_IncrementalMesh.hxx>
-#include <BRep_Tool.hxx>
-#include <GCPnts_TangentialDeflection.hxx>
-#include <NCollection_IndexedMap.hxx>
-#include <NCollection_Vec3.hxx>
-#include <Poly_Triangulation.hxx>
-#include <Standard_Failure.hxx>
-#include <TopAbs_Orientation.hxx>
-#include <TopExp.hxx>
-#include <TopExp_Explorer.hxx>
-#include <TopLoc_Location.hxx>
-#include <TopTools_ShapeMapHasher.hxx>
-#include <TopoDS.hxx>
-#include <TopoDS_Face.hxx>
-
-#include <cstdlib>
-#include <stdexcept>
-#include <string>
-
-MeshData OcctKernel::tessellate(uint32_t id, double linearDeflection, double angularDeflection) {
-    try {
-        const auto& shape = get(id);
-
-        BRepMesh_IncrementalMesh mesher(shape, linearDeflection, false, angularDeflection, false);
-        if (!mesher.IsDone()) {
-            throw std::runtime_error("tessellate: meshing failed");
-        }
-
-        // Count totals
-        int totalNodes = 0;
-        int totalTris = 0;
-        int totalFaces = 0;
-        for (TopExp_Explorer ex(shape, TopAbs_FACE); ex.More(); ex.Next()) {
-            TopLoc_Location loc;
-            auto tri = BRep_Tool::Triangulation(TopoDS::Face(ex.Current()), loc);
-            if (tri.IsNull())
-                continue;
-            totalNodes += tri->NbNodes();
-            totalTris += tri->NbTriangles();
-            totalFaces++;
-        }
-
-        MeshData result;
-        result.positionCount = totalNodes * 3;
-        result.normalCount = totalNodes * 3;
-        result.indexCount = totalTris * 3;
-
-        result.positions = static_cast<float*>(std::malloc(result.positionCount * sizeof(float)));
-        result.normals = static_cast<float*>(std::malloc(result.normalCount * sizeof(float)));
-        result.indices = static_cast<uint32_t*>(std::malloc(result.indexCount * sizeof(uint32_t)));
-        result.faceGroupCount = totalFaces * 3;
-        result.faceGroups =
-            static_cast<int32_t*>(std::malloc(result.faceGroupCount * sizeof(int32_t)));
-
-        if ((!result.positions && result.positionCount > 0) ||
-            (!result.normals && result.normalCount > 0) ||
-            (!result.indices && result.indexCount > 0) ||
-            (!result.faceGroups && result.faceGroupCount > 0)) {
-            throw std::runtime_error("tessellate: memory allocation failed");
-        }
-
-        int vertexOffset = 0;
-        int triOffset = 0;
-        int faceGroupIdx = 0;
-
-        for (TopExp_Explorer ex(shape, TopAbs_FACE); ex.More(); ex.Next()) {
-            const auto& face = TopoDS::Face(ex.Current());
-            TopLoc_Location loc;
-            auto tri = BRep_Tool::Triangulation(face, loc);
-            if (tri.IsNull())
-                continue;
-
-            const auto& trsf = loc.Transformation();
-            int nbNodes = tri->NbNodes();
-            int nbTri = tri->NbTriangles();
-
-            // Positions
-            for (int i = 1; i <= nbNodes; i++) {
-                gp_Pnt p = tri->Node(i).Transformed(trsf);
-                int base = (vertexOffset + i - 1) * 3;
-                result.positions[base + 0] = static_cast<float>(p.X());
-                result.positions[base + 1] = static_cast<float>(p.Y());
-                result.positions[base + 2] = static_cast<float>(p.Z());
-            }
-
-            // Normals
-            if (!tri->HasNormals()) {
-                BRepLib_ToolTriangulatedShape::ComputeNormals(face, tri);
-            }
-            for (int i = 1; i <= nbNodes; i++) {
-                gp_Dir d(0, 0, 1);
-                if (tri->HasNormals()) {
-                    NCollection_Vec3<float> nv;
-                    tri->Normal(i, nv);
-                    if (nv.x() != 0.0f || nv.y() != 0.0f || nv.z() != 0.0f) {
-                        d = gp_Dir(nv.x(), nv.y(), nv.z());
-                    }
-                }
-                d = d.Transformed(trsf);
-                int base = (vertexOffset + i - 1) * 3;
-                result.normals[base + 0] = static_cast<float>(d.X());
-                result.normals[base + 1] = static_cast<float>(d.Y());
-                result.normals[base + 2] = static_cast<float>(d.Z());
-            }
-
-            // Triangles (with winding correction for reversed faces)
-            bool isReversed = (face.Orientation() != TopAbs_FORWARD);
-            for (int t = 1; t <= nbTri; t++) {
-                const auto& triangle = tri->Triangle(t);
-                int n1 = triangle.Value(1);
-                int n2 = triangle.Value(2);
-                int n3 = triangle.Value(3);
-
-                if (isReversed) {
-                    int tmp = n1;
-                    n1 = n2;
-                    n2 = tmp;
-                }
-
-                result.indices[triOffset + 0] = static_cast<uint32_t>(n1 - 1 + vertexOffset);
-                result.indices[triOffset + 1] = static_cast<uint32_t>(n2 - 1 + vertexOffset);
-                result.indices[triOffset + 2] = static_cast<uint32_t>(n3 - 1 + vertexOffset);
-                triOffset += 3;
-            }
-
-            // Record face group: [triStart (in index units), triCount (indices), faceHash]
-            int faceTriStart = triOffset - nbTri * 3;
-            int faceHash = static_cast<int>(TopTools_ShapeMapHasher{}(face) % 2147483647);
-            result.faceGroups[faceGroupIdx + 0] = faceTriStart;
-            result.faceGroups[faceGroupIdx + 1] = nbTri * 3;
-            result.faceGroups[faceGroupIdx + 2] = faceHash;
-            faceGroupIdx += 3;
-
-            vertexOffset += nbNodes;
-        }
-
-        return result;
-    } catch (const Standard_Failure& e) {
-        throw std::runtime_error(std::string("tessellate: ") + e.what());
-    }
-}
-
-MeshData OcctKernel::meshShape(uint32_t id, double linearDeflection, double angularDeflection) {
-    return tessellate(id, linearDeflection, angularDeflection);
-}
-
-MeshBatchData OcctKernel::meshBatch(std::vector<uint32_t> ids, double linearDeflection,
-                                    double angularDeflection) {
-    try {
-        // First pass: mesh all shapes and count totals
-        struct ShapeMesh {
-            int posStart, posCount, idxStart, idxCount;
-        };
-        std::vector<ShapeMesh> shapeMeshes;
-        shapeMeshes.reserve(ids.size());
-
-        int totalNodes = 0;
-        int totalTris = 0;
-
-        for (uint32_t id : ids) {
-            const auto& shape = get(id);
-            BRepMesh_IncrementalMesh mesher(shape, linearDeflection, false, angularDeflection,
-                                            false);
-
-            int shapeNodes = 0;
-            int shapeTris = 0;
-            for (TopExp_Explorer ex(shape, TopAbs_FACE); ex.More(); ex.Next()) {
-                TopLoc_Location loc;
-                auto tri = BRep_Tool::Triangulation(TopoDS::Face(ex.Current()), loc);
-                if (tri.IsNull())
-                    continue;
-                shapeNodes += tri->NbNodes();
-                shapeTris += tri->NbTriangles();
-            }
-            shapeMeshes.push_back({totalNodes * 3, shapeNodes * 3, totalTris * 3, shapeTris * 3});
-            totalNodes += shapeNodes;
-            totalTris += shapeTris;
-        }
-
-        // Allocate
-        MeshBatchData result;
-        result.positionCount = totalNodes * 3;
-        result.normalCount = totalNodes * 3;
-        result.indexCount = totalTris * 3;
-        result.shapeCount = static_cast<int>(ids.size());
-
-        result.positions = static_cast<float*>(std::malloc(result.positionCount * sizeof(float)));
-        result.normals = static_cast<float*>(std::malloc(result.normalCount * sizeof(float)));
-        result.indices = static_cast<uint32_t*>(std::malloc(result.indexCount * sizeof(uint32_t)));
-        result.shapeOffsets =
-            static_cast<int32_t*>(std::malloc(result.shapeCount * 4 * sizeof(int32_t)));
-
-        // Second pass: extract geometry
-        int vertexOffset = 0;
-        int triOffset = 0;
-
-        for (size_t si = 0; si < ids.size(); si++) {
-            const auto& shape = get(ids[si]);
-
-            for (TopExp_Explorer ex(shape, TopAbs_FACE); ex.More(); ex.Next()) {
-                const auto& face = TopoDS::Face(ex.Current());
-                TopLoc_Location loc;
-                auto tri = BRep_Tool::Triangulation(face, loc);
-                if (tri.IsNull())
-                    continue;
-
-                const auto& trsf = loc.Transformation();
-                int nbNodes = tri->NbNodes();
-                int nbTri = tri->NbTriangles();
-
-                for (int i = 1; i <= nbNodes; i++) {
-                    gp_Pnt p = tri->Node(i).Transformed(trsf);
-                    int base = (vertexOffset + i - 1) * 3;
-                    result.positions[base + 0] = static_cast<float>(p.X());
-                    result.positions[base + 1] = static_cast<float>(p.Y());
-                    result.positions[base + 2] = static_cast<float>(p.Z());
-                }
-
-                if (!tri->HasNormals()) {
-                    BRepLib_ToolTriangulatedShape::ComputeNormals(face, tri);
-                }
-                for (int i = 1; i <= nbNodes; i++) {
-                    gp_Dir d(0, 0, 1);
-                    if (tri->HasNormals()) {
-                        NCollection_Vec3<float> nv;
-                        tri->Normal(i, nv);
-                        if (nv.x() != 0.0f || nv.y() != 0.0f || nv.z() != 0.0f) {
-                            d = gp_Dir(nv.x(), nv.y(), nv.z());
-                        }
-                    }
-                    d = d.Transformed(trsf);
-                    int base = (vertexOffset + i - 1) * 3;
-                    result.normals[base + 0] = static_cast<float>(d.X());
-                    result.normals[base + 1] = static_cast<float>(d.Y());
-                    result.normals[base + 2] = static_cast<float>(d.Z());
-                }
-
-                bool isReversed = (face.Orientation() != TopAbs_FORWARD);
-                for (int t = 1; t <= nbTri; t++) {
-                    const auto& triangle = tri->Triangle(t);
-                    int n1 = triangle.Value(1);
-                    int n2 = triangle.Value(2);
-                    int n3 = triangle.Value(3);
-                    if (isReversed)
-                        std::swap(n1, n2);
-                    result.indices[triOffset + 0] = static_cast<uint32_t>(n1 - 1 + vertexOffset);
-                    result.indices[triOffset + 1] = static_cast<uint32_t>(n2 - 1 + vertexOffset);
-                    result.indices[triOffset + 2] = static_cast<uint32_t>(n3 - 1 + vertexOffset);
-                    triOffset += 3;
-                }
-
-                vertexOffset += nbNodes;
-            }
-
-            // Store per-shape offsets: [posStart, posCount, idxStart, idxCount]
-            int oi = static_cast<int>(si) * 4;
-            result.shapeOffsets[oi + 0] = shapeMeshes[si].posStart;
-            result.shapeOffsets[oi + 1] = shapeMeshes[si].posCount;
-            result.shapeOffsets[oi + 2] = shapeMeshes[si].idxStart;
-            result.shapeOffsets[oi + 3] = shapeMeshes[si].idxCount;
-        }
-
-        return result;
-    } catch (const Standard_Failure& e) {
-        throw std::runtime_error(std::string("meshBatch: ") + e.what());
-    }
-}
-
-EdgeData OcctKernel::wireframe(uint32_t id, double deflection) {
-    try {
-        const auto& shape = get(id);
-
-        struct EdgeSample {
-            std::vector<gp_Pnt> pts;
-            int hash;
-        };
-        std::vector<EdgeSample> edgeSamples;
-        int totalPoints = 0;
-
-        // Use IndexedMap to avoid duplicate edges (shared between faces)
-        NCollection_IndexedMap<TopoDS_Shape, TopTools_ShapeMapHasher> edgeMap;
-        TopExp::MapShapes(shape, TopAbs_EDGE, edgeMap);
-        for (int ei = 1; ei <= edgeMap.Extent(); ei++) {
-            BRepAdaptor_Curve curve(TopoDS::Edge(edgeMap.FindKey(ei)));
-            GCPnts_TangentialDeflection sampler(curve, deflection, 0.5);
-            EdgeSample es;
-            for (int i = 1; i <= sampler.NbPoints(); i++) {
-                es.pts.push_back(sampler.Value(i));
-            }
-            es.hash = static_cast<int>(TopTools_ShapeMapHasher{}(edgeMap.FindKey(ei)) % 2147483647);
-            totalPoints += static_cast<int>(es.pts.size());
-            edgeSamples.push_back(std::move(es));
-        }
-
-        EdgeData result;
-        result.pointCount = totalPoints * 3;
-        result.points = static_cast<float*>(std::malloc(result.pointCount * sizeof(float)));
-        int numEdges = static_cast<int>(edgeSamples.size());
-        result.edgeGroupCount = numEdges * 3;
-        result.edgeGroups =
-            static_cast<int32_t*>(std::malloc(result.edgeGroupCount * sizeof(int32_t)));
-        if (!result.points && result.pointCount > 0) {
-            throw std::runtime_error("wireframe: allocation failed");
-        }
-
-        int offset = 0;
-        int edgeIdx = 0;
-        for (const auto& es : edgeSamples) {
-            int edgeStart = offset;
-            for (const auto& p : es.pts) {
-                result.points[offset + 0] = static_cast<float>(p.X());
-                result.points[offset + 1] = static_cast<float>(p.Y());
-                result.points[offset + 2] = static_cast<float>(p.Z());
-                offset += 3;
-            }
-            if (result.edgeGroups) {
-                result.edgeGroups[edgeIdx * 3] = edgeStart;
-                result.edgeGroups[edgeIdx * 3 + 1] = offset - edgeStart;
-                result.edgeGroups[edgeIdx * 3 + 2] = es.hash;
-            }
-            edgeIdx++;
-        }
-
-        return result;
-    } catch (const Standard_Failure& e) {
-        throw std::runtime_error(std::string("wireframe: ") + e.what());
-    }
-}
